+++ Lib/Watcher/OsWatcher.cs
@@ -4,17 +4,6 @@ using Lib.Utils;
 
 namespace Lib.Watcher
 {
-    public class DummyWatcher : IDirectoryWatcher
-    {
-        public void Dispose()
-        {
-        }
-
-        public string WatchedDirectory { get; set; }
-        public Action<string> OnFileChange { get; set; }
-        public Action OnError { get; set; }
-    }
-
     public class OsWatcher : IDirectoryWatcher
     {
         public string WatchedDirectory
@@ -102,8 +91,7 @@ namespace Lib.Watcher
             lock (_createLock)
             {
                 _disposed = true;
-                if (_fileSystemWatcher != null)
-                    _fileSystemWatcher.Dispose();
+                _fileSystemWatcher?.Dispose();
             }
         }
     }